[Feat] #57 - 모달 연결
